requirements.txt
accelerate==1.0.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
blobfile==3.0.0
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fairscale==0.4.13
filelock==3.16.1
fire==0.7.0
Flask==3.0.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.54.1
fsspec==2024.9.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
kiwisolver==1.4.7
llama_models==0.0.42
lxml==5.3.0
Mako==1.3.5
MarkupSafe==3.0.2
matplotlib==3.9.2
mock==5.1.0
mpmath==1.3.0
msgspec==0.18.6
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.3.3
openai==1.52.0
packaging==24.1
pandas==2.2.3
payment==0.0.1.2
pillow==11.0.0
plotly==5.24.1
psutil==6.1.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.4.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.46.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
WTForms==3.1.2
